Move library directory
