+++ lib/os/ffi/library/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "os"
+version = "0.1.0"
+publish = false
+edition = "2021"
+
+[lib]
+crate-type = ["staticlib"]
+
+[dependencies]
+ffi = { package = "pen-ffi", version = "0.3" }
+rand = "0.8"
+tempfile = "3"
