+++ lib/os/ffi/library/rust-toolchain.toml
@@ -1,0 +1,8 @@
+[toolchain]
+channel = "stable"
+targets = [
+  "i686-unknown-linux-musl",
+  "x86_64-unknown-linux-musl",
+  "aarch64-unknown-linux-musl",
+  "wasm32-wasi",
+]
