+++ lib/os/ffi/library/src/directory.rs
