+++ lib/os/ffi/library/src/environment_variable.rs
