+++ lib/os/ffi/library/src/error.rs
