+++ lib/os/ffi/library/src/result.rs
