+++ lib/os/ffi/library/src/string_array.rs
